lexer: fix bug that consumes input without handling it
Multi-character tokens usually require more careful parsing than their
single-letter counterparts. We have to make sure, for example, that the
subsequent char is part of a different token, and not the continuation
of the current one which is not really what we thought it were. The
lexer is not careful enough in such cases and it might end up forgetting
to reset the input pointer when it notices that the stream it was
analysing is not what it previously thought. Here is an example input
for which this bug happens:

	int main() { return0; }

In this case, the parser sees "return" and thinks it might be a return
keyword. Then it notices the adjacent "0" and realizes it is not the
return keyword, so it tries to "pattern match" the stream to another
token type. However, it had already consumed the "return" part, so "0"
fits for an integer token. Instead, it should rewind and keep trying to
match the whole "return0" sequence, until it finally notices that this
is an invalid stream, which cannot be matched to any known token. (In
the future we probably want to allow this as a valid identifier, but we
currently only allow alphabetical chars in identifiers.) Let's fix this
problem by advancing a different pointer and only updating our input
pointer when we know for sure that the analysed data is a valid token.