parser: support stage 2 tokens (-, !, and ~)
This is the first case that one of our parser functions do not expect a
fixed sequence of tokens, but instead can process any one of a list. To
make room for this change, replace the simple expect_type() macro by a
more complex function that takes a list of token types and checks
whether the next token on the input stream corresponds to any type on
the list. While we are here, let's also rename and modify the macro's
usage to be more explicit about its destructive behavior on the `tok`
parameter.