lexer: add stage-3 tokens (+, *, and /)
This will correspond to the following binary operations: addition,
multiplication, and division. We already have '-' (TOK_MINUS) for the
subtraction and the parentheses tokens to express precedence.

Note: we are using TOK_MINUS, TOK_STAR, etc., instead of
TOK_SUBTRACTION, and TOK_MULTIPLICATION, because we might want to use
these tokens to represent other AST nodes in the future. For example,
TOK_STAR might be also used for pointer dereferencing.