lexer, parser: when showing source line on errors, replace '\t' by ' '
We don't know which width will be used by the terminal to display tabs,
and this can confuse our column counter. For example, if we had a source
line with "\treturn + 5;", the parser would currently produce the
following message:

fatal: parser: expecting <integer>, (, -, ~, or ! got +
On line 2:      return + 5;
                   ^

Note that the '^' indicator is at the wrong column because the tab ended
up taking more columns than what we accounted for (which was one). Let's
fix that by replacing each tab with a single whitespace.