x86: support stage-4 logic operators
Note: I implemented the logical AND and OR instructions a little
different than the blog post I'm following[1]. The idea was to use less
instructions and make less jmps. But I'm not sure if the performance is
actually better because, when we skip the second clause, the number os
instructions is actually one higher. Nevertheless, if I ever choose to
replace it with the version from the blog, the code should be the
following:

static char *label_or_2nd_clause(void)
{
	static unsigned long counter = 0;
	return xmkstr("_or_2nd_clause_%lu", counter++);
}

static void generate_logic_or(struct ast_expression *exp, FILE *file)
{
	assert(exp->type == AST_EXP_BINARY_OP &&
	       exp->u.bin_op.type == EXP_OP_LOGIC_OR);

	char *label_2nd_clause = label_or_2nd_clause();
	char *label_skip_2nd_clause = label_or_skip_2nd_clause();

	generate_expression(exp->u.bin_op.lexp, file);
	xfprintf(file, " cmp	$0, %%eax\n");
	xfprintf(file, " je	%s\n", label_2nd_clause);
	xfprintf(file, " mov	$1, %%eax\n"); /* lexp is true, set result to 1 */
	xfprintf(file, " jmp	%s\n", label_skip_2nd_clause);
	xfprintf(file, "%s:\n", label_2nd_clause);
	xfprintf(file, " cmp	$0, %%eax\n");
	generate_expression(exp->u.bin_op.rexp, file);
	xfprintf(file, " cmp	$0, %%eax\n");
	xfprintf(file, " mov	$0, %%eax\n");
	xfprintf(file, " setne	%%al\n");
	xfprintf(file, "%s:\n", label_skip_2nd_clause);
	free(label_2nd_clause);
	free(label_skip_2nd_clause);
}

[1]: https://norasandler.com/2017/12/28/Write-a-Compiler-4.html#short-circuit-evaluation