parser: make assignment a binary operator
This not only simplifies the code but also makes it easier to reason
about the precedence and associativity of the assignment operator. The
only downside is that we no longer enforce that the left operand must be
an lvalue in the data struct itself, but we still do it in the code.