lib: rename var_map to strmap and make it more generic
Looking to the next blog posts in the series, I saw that we will need to
handle other symbols besides local variables, like global variables and
functions. While in time, I want to make it easier to add those symbols
in the future by constructing a sym_map, which will be built upon this
generic API (the strmap).