x86: code generation for stage_5: local variables
There are 3 major code generation changes at this stage:

- Multiple statement functions.
- Statements can be expressions and variable declarations (besides
  return).
- Functions may not have a defined return statement. In this case,
  main() should return 0.
- Generate function prologue and epilogue, to keep track of %rbp, which
  is how we reference local variable's addresses.
- Create code for variable assignment and referencing.

Furthermore, although the variable mapping could be implemented with
only a hashtable, I decided to use a full symbol mapping structure to
support further changes in the upcoming stages. This should help with
the addition of global variables and other functions. All these symbols
can be stored in the same symtable and the symtable routines can be
taught to detect errors such as symbol redefinition at a different type.
For the shadowing problem, my idea is to use a "unsigned long" scope id,
which is incremented every time we enter a block, and assign this id to
the sym_data on a symbol declaration. Then, we can check whether a new
declaration of the same symbol shadows a symbol from a previous scope
(which is fine) or redefines a symbol of the same scope (which is an
error).

Finally, note that the stack_index is currently always incremented by 8,
which is the size of the registers we push to the stack. I'm not sure
yet if I should be pushing 4 (bytes) instead... I thought for x86_64 we
wanted to push 8, but this deserves more investigation... Also, I guess
that if, one day, we have other types with different sizes, the
increment size will be different for each of them.