x86: local var declaration: fix weird case of "self assignment"
Previously, when generating code for a local variable declaration, we
would first generate the code for the rexp, then put the variable's name
into our compiler-time symbol table. However, the expression might
itself reference the variable in this weird case:

	int v = v;

This doesn't seem very practical, but since gcc allows it, let's allow
it too. To do that, we just need to put "v" in the symbol table before
generating rexp.

Note: we only need to make this change on variable declaration, but
let's replicate it on variable assignment for consistency.