parser: support prefix and suffix inc/dec ops (++ and --)
Unlink compound assignment operators, we will implement the increment
and decrement operators as a new expression type. Perhaps they could be
a syntax sugar for already existing operators, but it seems more complex
to do that than what it was for the compound assignment. Also (but not
as important for this project), handling them directly might allow us to
produce a more optimized assembly.