parser: disallow variable declaration at if-else statements
The goal of this commit it to properly error out when we find something
like this:

    if (...)
        int x = 2;
    else
        int y;

Which is not allowed by the C standard. In fact, the grammar[1] given
with the standard even says that variable declarations are not
statements. And the tutorial blob I'm following also removes it from the
statement definition at stage 6. However, this would require a pretty
big change and, in the hopes of simplification, I'm using enforcing this
rule in the parser code, rather than in the struct definitions. Let's
see later if this turns out to be a good decision or if I should indeed
move declarations to a separate struct.

[1]: http://www.open-std.org/jtc1/sc22/wg14/www/docs/n1570.pdf#page=476