x86: generate code for function calls
Note: we do a little hack changing the type of
'struct ast_func_decl'.parameters to an array of
'struct ast_var_decl *'. This is done to assist the later calls to
symtable_put_lvar(). Also note that no validation is implemented yet, so
the compiler still does not error out on things like argument number
mismatch.