x86: use only 4 bytes to store ints on stack (local vars)
Previously we uses 8 bytes as we cannot push 4 bytes onto the stack on
x86-64. This patches does a little hack to emulate the 4-bytes push by
first subtracting 4 bytes from %rsp (to "allocate" space) and them
moving the value to it. I'm not very confident with assembly and I don't
know if this design might turn out to be problematic later (e.g. due to
stack alignment issues). Let's see.

Note that, for function calls, we still have to push 8 bytes into the
stack, even if we are only passing 4-byte integers. That's how the ABI
works.